Animals never follow player if they throw food
